+++ version/changelog.md
@@ -10,6 +10,7 @@
 - typing text will execute commands in cmdline
 - delete key should also delete file like F8
 - enable sorting
+- rename files using F2
 
 ### v1.1.0 (26.01.2026)